Clean up ServerNode interface
Remove unnecessary overloads and add javadoc to reference-related methods.

+++ milo-examples/server-examples/src/main/java/org/eclipse/milo/examples/server/ExampleNamespace.java
@@ -509,7 +509,7 @@ public class ExampleNamespace implements Namespace {
             if (node != null) {
                 DataValue value = node.readAttribute(
                     new AttributeContext(context),
-                    readValueId.getAttributeId().intValue(),
+                    readValueId.getAttributeId(),
                     timestamps,
                     readValueId.getIndexRange()
                 );
