+++ opc-ua-sdk/sdk-server/src/main/java/org/eclipse/milo/opcua/sdk/server/namespaces/OpcUaNamespace.java
@@ -127,7 +127,7 @@ public class OpcUaNamespace implements Namespace {
             if (node != null) {
                 value = node.readAttribute(
                     new AttributeContext(context),
-                    id.getAttributeId().intValue(),
+                    id.getAttributeId(),
                     timestamps,
                     id.getIndexRange()
                 );
